Make classDirectory+dependencyClasspath configurable
As mentioned in [#63](https://github.com/ktoso/sbt-jmh/issues/63#issuecomment-230509786)
one can hack around the hardcoded config that is used for `classDirectory`
and `dependencyClasspath` (both are taken from the `Compile` config).

This change allows the user to set

```
classDirectory in Jmh := (classDirectory in Test).value
dependencyClasspath in Jmh := (dependencyClasspath in Test).value
```

so that jmh will use the appropriate values.