Disable BSP in the Jmh config
Importing a Jmh build target in IntelliJ or Metals triggers
the Jmh source generation which triggers the compilation.
This is not user friendly because:
- it takes time to import the project
- it can compile in error and then it makes the entire import fail
(we need more granularity in BSP to handle individual failures)

Importing the Jmh build target is often not useful because it shares
its sources with the Compile or Test build target.
So we disable BSP in the Jmh config by default.